Add SpotifyTable component
Removed code repetition related to tables, now there is one Table component which behaves differently based on the type props. Single and double click are now working on every table.
