+++ frontend/src/components/main/Album.tsx
@@ -2,12 +2,9 @@ import React, {FC, useEffect, useState} from "react";
 import {Link, useParams} from "react-router-dom";
 import AxiosClient from "../../utils/axiosClient";
 import {SpotifySimplifiedTrackObject} from "../../types/spotify";
-import {Grid, Table, TableBody, TableCell, TableContainer, TableHead, TableRow} from "@material-ui/core";
-import {getMsToTime, getPlaylistLength} from "../../utils/dataFormat";
-import InfiniteScroll from "react-infinite-scroll-component";
-import AccessTimeIcon from "@material-ui/icons/AccessTime";
-import {getArtistsWithLinks} from "../../utils/formatComponents";
-import ExplicitIcon from '@material-ui/icons/Explicit';
+import {Grid} from "@material-ui/core";
+import {getPlaylistLength} from "../../utils/dataFormat";
+import SpotifyTable from "../layout/SpotifyTable";
 
 interface Parameters {
   id: string,
@@ -69,48 +66,12 @@ const Album: FC = () => {
       </Grid>}
 
       <Grid item xs={12} className="playlist__tracks">
-        <TableContainer>
-          <InfiniteScroll
-            next={loadMoreTracks}
-            hasMore={next != null}
-            loader={<h2>Loading more tracks ...</h2>}
-            dataLength={tracks.length}
-            scrollableTarget='content'
-          >
-            <Table aria-label="simple table" size="small">
-              <TableHead className="playlist__tracks-header">
-                <TableRow>
-                  <TableCell align="left">#</TableCell>
-                  <TableCell align="left">TITLE</TableCell>
-                  <TableCell align="left">REPLAYS</TableCell>
-                  <TableCell align="left"><AccessTimeIcon fontSize="small"/></TableCell>
-                </TableRow>
-              </TableHead>
-              <TableBody>
-                {tracks.map(
-                  ({artists, duration_ms, name, explicit}, index) => (
-                    <TableRow key={`track-${index}`} className="playlist__track-row">
-                      <TableCell component="th" scope="row">{index + 1}</TableCell>
-                      <TableCell align="left">
-                        <Grid container alignItems="center">
-                          <Grid item>
-                            <Grid item className="playlist__track-title">
-                              <span>{name}</span>
-                            </Grid>
-                            <Grid item className="playlist__track-artists">
-                              {explicit && <ExplicitIcon/>}{getArtistsWithLinks(artists)}
-                            </Grid>
-                          </Grid>
-                        </Grid>
-                      </TableCell>
-                      <TableCell align="left">{""}</TableCell>
-                      <TableCell align="left">{getMsToTime(duration_ms, true)}</TableCell>
-                    </TableRow>
-                  ))}
-              </TableBody>
-            </Table>
-          </InfiniteScroll>
-        </TableContainer>
+        <SpotifyTable
+          tableType="album"
+          tracks={tracks}
+          next={next}
+          loadMore={loadMoreTracks}
+        />
       </Grid>
     </div>
   );
